changed datacount size to uint16 in ethernet and serial

+++ src/RoveCommSerial/RoveCommSerial.cpp
@@ -23,7 +23,7 @@ struct rovecomm_packet RoveCommSerial::read()
     //Todo: Why don't we pass these directly into the rovecomm_packet we just made?
     uint16_t data_id    =  0;
     roveware::data_type_t data_type;
-    uint8_t data_count =  0;
+    uint16_t data_count =  0;
           
 	//Create array to take packet
     uint8_t _packet[255];
@@ -45,7 +45,7 @@ struct rovecomm_packet RoveCommSerial::read()
 }
 
 ///////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
-void RoveCommSerial::_write(const uint8_t data_type_length, const roveware::data_type_t data_type, const uint16_t data_id, const uint8_t data_count, const void* data)
+void RoveCommSerial::_write(const uint8_t data_type_length, const roveware::data_type_t data_type, const uint16_t data_id, const uint16_t data_count, const void* data)
 { 
   //Creat packed udp packet
   struct roveware::_packet _packet = roveware::packPacket(data_id, data_count, data_type, data);
@@ -62,32 +62,32 @@ void RoveCommSerial::_write(const uint8_t data_type_length, const roveware::data
 //                   data_p[0] = data;
 //                   this->_write( 4,  roveware::INT32_T, data_id,               data_count,        (void*) data_p ); }
 //
-void RoveCommSerial::write(        const uint16_t  data_id, const uint8_t data_count, const  int32_t data )
+void RoveCommSerial::write(        const uint16_t  data_id, const uint16_t data_count, const  int32_t data )
 {                  int32_t data_p[1];
                    data_p[0] = data;
                    this->_write( 4,  roveware::INT32_T, data_id,               data_count,        (void*) data_p ); }
 
-void RoveCommSerial::write(        const uint16_t  data_id, const uint8_t data_count, const uint32_t data )
+void RoveCommSerial::write(        const uint16_t  data_id, const uint16_t data_count, const uint32_t data )
 {                  uint32_t data_p[1];
                    data_p[0] = data;
                    this->_write( 4, roveware::UINT32_T, data_id,               data_count,        (void*) data_p ); }
 
-void RoveCommSerial::write(        const uint16_t  data_id, const uint8_t data_count, const  int16_t data )
+void RoveCommSerial::write(        const uint16_t  data_id, const uint16_t data_count, const  int16_t data )
 {                  int16_t data_p[1];
                    data_p[0] = data;
                    this->_write( 2,  roveware::INT16_T, data_id,               data_count,        (void*) data_p ); }
 
-void RoveCommSerial::write(        const uint16_t  data_id, const uint8_t data_count, const uint16_t data )
+void RoveCommSerial::write(        const uint16_t  data_id, const uint16_t data_count, const uint16_t data )
 {                  uint16_t data_p[1];
                    data_p[0] = data;
                    this->_write( 2, roveware::UINT16_T, data_id,               data_count,        (void*) data_p ); }
 
-void RoveCommSerial::write(         const uint16_t data_id, const uint8_t data_count, const   int8_t data )
+void RoveCommSerial::write(         const uint16_t data_id, const uint16_t data_count, const   int8_t data )
 {                  int8_t data_p[1];
                    data_p[0] = data;
                    this->_write( 1,   roveware::INT8_T, data_id,               data_count,        (void*) data_p ); }
 
-void RoveCommSerial::write(        const uint16_t  data_id, const uint8_t data_count, const  uint8_t data )
+void RoveCommSerial::write(        const uint16_t  data_id, const uint16_t data_count, const  uint8_t data )
 {                  uint8_t data_p[1];
                    data_p[0] = data;
                    this->_write( 1,  roveware::UINT8_T, data_id,               data_count,        (void*) data_p ); }
@@ -95,20 +95,20 @@ void RoveCommSerial::write(        const uint16_t  data_id, const uint8_t data_c
 //void RoveCommSerial::write(        const  int      data_id, const  int    data_count, const  int     *data ) 
 //{                  this->_write( 4,  roveware::INT32_T, data_id,               data_count,        (void*) data ); }
 //
-void RoveCommSerial::write(        const uint16_t  data_id, const uint8_t data_count, const  int32_t *data )
+void RoveCommSerial::write(        const uint16_t  data_id, const uint16_t data_count, const  int32_t *data )
 {                  this->_write( 4,  roveware::INT32_T, data_id,               data_count,        (void*) data ); }
 
-void RoveCommSerial::write(        const uint16_t  data_id, const uint8_t data_count, const uint32_t *data )
+void RoveCommSerial::write(        const uint16_t  data_id, const uint16_t data_count, const uint32_t *data )
 {                  this->_write( 4, roveware::UINT32_T, data_id,               data_count,        (void*) data ); }
 
-void RoveCommSerial::write(        const uint16_t  data_id, const uint8_t data_count, const  int16_t *data )
+void RoveCommSerial::write(        const uint16_t  data_id, const uint16_t data_count, const  int16_t *data )
 {                  this->_write( 2,  roveware::INT16_T, data_id,               data_count,        (void*) data ); }
 
-void RoveCommSerial::write(        const uint16_t  data_id, const uint8_t data_count, const uint16_t *data )
+void RoveCommSerial::write(        const uint16_t  data_id, const uint16_t data_count, const uint16_t *data )
 {                  this->_write( 2, roveware::UINT16_T, data_id,               data_count,        (void*) data ); }
 
-void RoveCommSerial::write(         const uint16_t data_id, const uint8_t data_count, const   int8_t *data )
+void RoveCommSerial::write(         const uint16_t data_id, const uint16_t data_count, const   int8_t *data )
 {                  this->_write( 1,   roveware::INT8_T, data_id,               data_count,        (void*) data ); }
 
-void RoveCommSerial::write(        const uint16_t  data_id, const uint8_t data_count, const  uint8_t *data )
+void RoveCommSerial::write(        const uint16_t  data_id, const uint16_t data_count, const  uint8_t *data )
 {                  this->_write( 1,  roveware::UINT8_T, data_id,               data_count,        (void*) data ); }
