+++ src/RoveCommSerial/RoveCommSerial.h
@@ -16,7 +16,7 @@ class RoveCommSerial
 	
     //Called by overloaded write functions
     void _write(  const uint8_t  data_type_length, const roveware::data_type_t data_type, 
-                  const uint16_t data_id,    const uint8_t data_count, const void* data);
+                  const uint16_t data_id,    const uint16_t data_count, const void* data);
 
   public:
     /////begin///////////////////////////
@@ -30,22 +30,22 @@ class RoveCommSerial
 	//Overloaded for each data type
     //Causes bug when doing:RoveComm.write(SINGLE_VALUE_EXAMPLE_ID, 1, analogRead(A0));
 	//void write(const uint16_t data_id, const int     data_count, const int      data);
-    void write(const uint16_t data_id, const uint8_t data_count, const uint8_t  data);
-    void write(const uint16_t data_id, const uint8_t data_count, const uint16_t data);
-    void write(const uint16_t data_id, const uint8_t data_count, const uint32_t data);
-    void write(const uint16_t data_id, const uint8_t data_count, const int8_t   data);
-    void write(const uint16_t data_id, const uint8_t data_count, const int16_t  data);
-    void write(const uint16_t data_id, const uint8_t data_count, const int32_t  data);
+    void write(const uint16_t data_id, const uint16_t data_count, const uint8_t  data);
+    void write(const uint16_t data_id, const uint16_t data_count, const uint16_t data);
+    void write(const uint16_t data_id, const uint16_t data_count, const uint32_t data);
+    void write(const uint16_t data_id, const uint16_t data_count, const int8_t   data);
+    void write(const uint16_t data_id, const uint16_t data_count, const int16_t  data);
+    void write(const uint16_t data_id, const uint16_t data_count, const int32_t  data);
 
     //Array entry write
 	//Overloaded for each data type
     void write(const uint16_t data_id, const int     data_count, const int      *data);
-    void write(const uint16_t data_id, const uint8_t data_count, const uint8_t  *data);
-    void write(const uint16_t data_id, const uint8_t data_count, const uint16_t *data);
-    void write(const uint16_t data_id, const uint8_t data_count, const uint32_t *data);
-    void write(const uint16_t data_id, const uint8_t data_count, const int8_t   *data);
-    void write(const uint16_t data_id, const uint8_t data_count, const int16_t  *data);
-    void write(const uint16_t data_id, const uint8_t data_count, const int32_t  *data);
+    void write(const uint16_t data_id, const uint16_t data_count, const uint8_t  *data);
+    void write(const uint16_t data_id, const uint16_t data_count, const uint16_t *data);
+    void write(const uint16_t data_id, const uint16_t data_count, const uint32_t *data);
+    void write(const uint16_t data_id, const uint16_t data_count, const int8_t   *data);
+    void write(const uint16_t data_id, const uint16_t data_count, const int16_t  *data);
+    void write(const uint16_t data_id, const uint16_t data_count, const int32_t  *data);
   
 };
